backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.0
openai>=1.12.0 